Compute simulated (sample) average treatment effects on the treated from a Zelig model estimation

Compute simulated (sample) average treatment effects on the treated from a Zelig model estimation

ATT(object, treatment, treated = 1, num = NULL)

Arguments

object
an object of class Zelig

treatment
character string naming the variable that denotes the treatment and non-treated groups.

treated
value of treatment variable indicating treatment

num
number of simulations to run. Default is 1000.

Examples

library(dplyr)
data(sanction)
z.att <- zelig(num ~ target + coop + mil, model = "poisson",
                 data = sanction) %>%
             ATT(treatment = "mil") %>%
             get_qi(qi = "ATT", xvalue = "TE")

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   poisson: Poisson Regression for Event Count Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/