Zelig reference class

Zelig website: http://zeligproject.org/

Fields

fn

R function to call to wrap

formula

Zelig formula

weights

[forthcoming]

name

name of the Zelig model

data

data frame or matrix

by

split the data by factors

mi

work with imputed dataset

idx

model index

zelig.call

Zelig function call

model.call

wrapped function call

zelig.out

estimated zelig model(s)

setx.out

set values

setx.labels

pretty-print qi

bsetx

is x set?

bsetx1

is x1 set?

bsetrange

is range set?

bsetrange1

is range1 set?

range

range

range1

range1

test.statistics

list of test statistics

sim.out

simulated qi's

simparam

simulated parameters

num

number of simulations

authors

Zelig model authors

zeligauthors

Zelig authors

modelauthors

wrapped model authors

packageauthors

wrapped package authors

refs

citation information

year

model is released

description

model description

url

model URL

url.docs

model documentation URL

category

model category

vignette.url

vignette URL

json

JSON export

ljson

JSON export

outcome

JSON export

wrapper

JSON export

explanatory

JSON export

mcunit.test

unit testing

with.feedback

Feedback

Methods

ATT(treatment, treated = 1, quietly = TRUE, num = NULL)

Generic Method for Computing Simulated (Sample) Average Treatment Effects on the Treated

cite()

Provide citation information about Zelig and Zelig model, and about wrapped package and wrapped model

feedback()

Send feedback to the Zelig team

from_zelig_model()

Extract the original fitted model object from a zelig call. Note only works for models using directly wrapped functions.

get_coef()

Get estimated model coefficients

get_df_residual()

Get residual degrees-of-freedom

get_fitted()

Get estimated fitted values

get_model_data()

Get data used to estimate the model

get_names()

Return Zelig object field names

get_predict()

Get predicted values

get_pvalue()

Get estimated model p-values

get_qi(qi = "ev", xvalue = "x", subset = NULL)

Get quantities of interest

get_residuals()

Get estimated model residuals

get_se()

Get estimated model standard errors

get_vcov()

Get estimated model variance-covariance matrix

graph(...)

Plot the quantities of interest

help()

Open the model vignette from http://zeligproject.org/

packagename()

Automatically retrieve wrapped package name

references(style = "sphinx")

Construct a reference list specific to a Zelig model.

set(..., fn = list(numeric = mean, ordered = Median))

Setting Explanatory Variable Values

sim(num = NULL)

Generic Method for Computing and Organizing Simulated Quantities of Interest

simATT(simparam, data, depvar, treatment, treated)

Simulate an Average Treatment on the Treated

summarise(...)

Display a Zelig object

summarize(...)

Display a Zelig object

toJSON()

Convert Zelig object to JSON format

zelig(formula, data, model = NULL, ..., weights = NULL, by, bootstrap = FALSE)

The zelig function estimates a variety of statistical models