Time-Series Model with Moving Average

Warning: Simulation is not currently implemented for timeseries models.

Arguments

formula
a symbolic representation of the model to be estimated, in the form y \~\, x1 + x2, where y is the dependent variable and x1 and x2 are the explanatory variables, and y, x1, and x2 are contained in the same dataset. (You may include more than two explanatory variables, of course.) The + symbol means ``inclusion'' not ``addition.'' You may also include interaction terms and main effects in the form x1*x2 without computing them in prior steps; I(x1*x2) to include only the interaction term and exclude the main effects; and quadratic terms in the form I(x1^2).

model
the name of a statistical model to estimate. For a list of supported models and their documentation see: http://docs.zeligproject.org/articles/.

data
the name of a data frame containing the variables referenced in the formula or a list of multiply imputed data frames each having the same variable names and row numbers (created by Amelia or to_zelig_mi).

...
additional arguments passed to zelig, relevant for the model to be estimated.

by
a factor variable contained in data. If supplied, zelig will subset the data frame based on the levels in the by variable, and estimate a model for each subset. This can save a considerable amount of effort. For example, to run the same model on all fifty states, you could use: z.out <- zelig(y ~ x1 + x2, data = mydata, model = 'ls', by = 'state') You may also use by to run models using MatchIt subclasses.

cite
If is set to 'TRUE' (default), the model citation will be printed to the console.

ts
The name of the variable containing the time indicator. This should be passed in as a string. If this variable is not provided, Zelig will assume that the data is already ordered by time.

cs
Name of a variable that denotes the cross-sectional element of the data, for example, country name in a dataset with time-series across different countries. As a variable name, this should be in quotes. If this is not provided, Zelig will assume that all observations come from the same unit over time, and should be pooled, but if provided, individual models will be run in each cross-section. If cs is given as an argument, ts must also be provided. Additionally, by must be NULL.

order
A vector of length 3 passed in as c(p,d,q) where p represents the order of the autoregressive model, d represents the number of differences taken in the model, and q represents the order of the moving average model.

Value

Depending on the class of model selected, zelig will return an object with elements including coefficients, residuals, and formula which may be summarized using summary(z.out) or individually extracted using, for example, coef(z.out). See http://docs.zeligproject.org/articles/getters.html for a list of functions to extract model components. You can also extract whole fitted model objects using from_zelig_model.

Details

Currently only the Reference class syntax for time series. This model does not accept Bootstraps or weights.

See also

Vignette: http://docs.zeligproject.org/articles/zelig_ma.html

Examples

data(seatshare)
subset <- seatshare[seatshare$country == "UNITED KINGDOM",]
ts.out <- zma$new()
ts.out$zelig(formula = unemp ~ leftseat,  ts = "year", data = subset)
summary(ts.out)

#> Model: 
#> 
#> Call:
#> ts.out$zelig(formula = unemp ~ leftseat, data = subset, ts = "year", order = c(0, 
#> 0, 1))
#> 
#> Coefficients:
#>         ma1  intercept  leftseat
#>       1.000       10.6    -11.49
#> s.e.  0.066        1.9      3.99
#> 
#> sigma^2 estimated as 2.44:  log likelihood = -74.6,  aic = 157
#> Next step: Use 'setx' method