Bayesian Normal Linear Regression

Bayesian Normal Linear Regression

Arguments

formula
a symbolic representation of the model to be estimated, in the form y ~ x1 + x2, where y is the dependent variable and x1 and x2 are the explanatory variables, and y, x1, and x2 are contained in the same dataset. (You may include more than two explanatory variables, of course.) The + symbol means ``inclusion'' not ``addition.'' You may also include interaction terms and main effects in the form x1*x2 without computing them in prior steps; I(x1*x2) to include only the interaction term and exclude the main effects; and quadratic terms in the form I(x1^2).

model
the name of a statistical model to estimate. For a list of other supported models and their documentation see: http://docs.zeligproject.org/articles/.

data
the name of a data frame containing the variables referenced in the formula or a list of multiply imputed data frames each having the same variable names and row numbers (created by Amelia or to_zelig_mi).

...
additional arguments passed to zelig, relevant for the model to be estimated.

by
a factor variable contained in data. If supplied, zelig will subset the data frame based on the levels in the by variable, and estimate a model for each subset. This can save a considerable amount of effort. You may also use by to run models using MatchIt subclasses.

cite
If is set to 'TRUE' (default), the model citation will be printed to the console.

Value

Depending on the class of model selected, zelig will return an object with elements including coefficients, residuals, and formula which may be summarized using summary(z.out) or individually extracted using, for example, coef(z.out). See http://docs.zeligproject.org/articles/getters.html for a list of functions to extract model components. You can also extract whole fitted model objects using from_zelig_model.

Details

Additional parameters avaialable to many models include:

weights: vector of weight values or a name of a variable in the dataset by which to weight the model. For more information see: http://docs.zeligproject.org/articles/weights.html.

burnin: number of the initial MCMC iterations to be discarded (defaults to 1,000).

mcmc: number of the MCMC iterations after burnin (defaults to 10,000).

thin: thinning interval for the Markov chain. Only every thin-th draw from the Markov chain is kept. The value of mcmc must be divisible by this value. The default value is 1.

verbose: defaults to FALSE. If TRUE, the progress of the sampler (every 10%) is printed to the screen.

seed: seed for the random number generator. The default is NA which corresponds to a random seed of 12345.

beta.start: starting values for the Markov chain, either a scalar or vector with length equal to the number of estimated coefficients. The default is NA, such that the maximum likelihood estimates are used as the starting values.

Use the following parameters to specify the model's priors:

b0: prior mean for the coefficients, either a numeric vector or a scalar. If a scalar value, that value will be the prior mean for all the coefficients. The default is 0.

B0: prior precision parameter for the coefficients, either a square matrix (with the dimensions equal to the number of the coefficients) or a scalar. If a scalar value, that value times an identity matrix will be the prior precision parameter. The default is 0, which leads to an improper prior.

c0: c0/2 is the shape parameter for the Inverse Gamma prior on the variance of the disturbance terms.

d0: d0/2 is the scale parameter for the Inverse Gamma prior on the variance of the disturbance terms.

See also

Vignette: http://docs.zeligproject.org/articles/zelig_normalbayes.html

Examples

data(macro)
z.out <- zelig(unem ~ gdp + capmob + trade, model = "normal.bayes", data = macro, verbose = FALSE)

#> How to cite this model in Zelig:
#>   Ben Goodrich, and Ying Lu. 2013.
#>   normal-bayes: Bayesian Normal Linear Regression
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

data(macro)
z.out <- zelig(unem ~ gdp + capmob + trade, model = "normal.bayes",
data = macro, verbose = FALSE)

#> How to cite this model in Zelig:
#>   Ben Goodrich, and Ying Lu. 2013.
#>   normal-bayes: Bayesian Normal Linear Regression
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

z.out$geweke.diag()

#> 
#> Fraction in 1st window = 0.1
#> Fraction in 2nd window = 0.5 
#> 
#> (Intercept)         gdp      capmob       trade      sigma2 
#>      -0.118      -0.313      -0.589       0.516      -1.696 
#> 

z.out$heidel.diag()

#>                                           
#>             Stationarity start     p-value
#>             test         iteration        
#> (Intercept) passed       1         0.756  
#> gdp         passed       1         0.963  
#> capmob      passed       1         0.355  
#> trade       passed       1         0.339  
#> sigma2      passed       1         0.626  
#>                                        
#>             Halfwidth Mean    Halfwidth
#>             test                       
#> (Intercept) passed     6.1773 0.008849 
#> gdp         passed    -0.3240 0.001244 
#> capmob      passed     1.4207 0.003246 
#> trade       passed     0.0199 0.000105 
#> sigma2      passed     7.5834 0.011691 

z.out$raftery.diag()

#> 
#> Quantile (q) = 0.025
#> Accuracy (r) = +/- 0.005
#> Probability (s) = 0.95 
#>                                                    
#>              Burn-in  Total Lower bound  Dependence
#>              (M)      (N)   (Nmin)       factor (I)
#>  (Intercept) 2        3834  3746         1.020     
#>  gdp         2        3650  3746         0.974     
#>  capmob      2        3771  3746         1.010     
#>  trade       2        3680  3746         0.982     
#>  sigma2      2        3710  3746         0.990     
#> 

summary(z.out)

#> Model: 
#> 
#> Iterations = 1001:11000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 10000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                Mean      SD Naive SE Time-series SE
#> (Intercept)  6.1773 0.45148 4.51e-03       4.51e-03
#> gdp         -0.3240 0.06349 6.35e-04       6.35e-04
#> capmob       1.4207 0.16562 1.66e-03       1.66e-03
#> trade        0.0199 0.00561 5.61e-05       5.37e-05
#> sigma2       7.5834 0.57846 5.78e-03       5.96e-03
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%     25%    50%     75%   97.5%
#> (Intercept)  5.30927  5.8707  6.180  6.4779  7.0878
#> gdp         -0.45044 -0.3669 -0.324 -0.2812 -0.2012
#> capmob       1.09273  1.3105  1.422  1.5335  1.7461
#> trade        0.00873  0.0162  0.020  0.0237  0.0308
#> sigma2       6.53464  7.1766  7.553  7.9552  8.7957
#> 
#> Next step: Use 'setx' method

x.out <- setx(z.out)
s.out1 <- sim(z.out, x = x.out)
summary(s.out1)

#> 
#>  sim x :
#>  -----
#> ev
#>   mean    sd  50% 2.5% 97.5%
#> 1 4.99 0.148 4.99 4.71  5.28
#> pv
#>      mean   sd  50%   2.5% 97.5%
#> [1,] 5.02 2.79 5.04 -0.453  10.5

x.high <- setx(z.out, trade = quantile(macro$trade, prob = 0.8))
x.low <- setx(z.out, trade = quantile(macro$trade, prob = 0.2))

s.out2 <- sim(z.out, x = x.high, x1 = x.low)
summary(s.out2)

#> 
#>  sim x :
#>  -----
#> ev
#>   mean    sd  50% 2.5% 97.5%
#> 1 5.43 0.192 5.43 5.06  5.81
#> pv
#>      mean   sd  50%    2.5% 97.5%
#> [1,] 5.45 2.77 5.45 -0.0369  10.8
#> 
#>  sim x1 :
#>  -----
#> ev
#>   mean    sd 50% 2.5% 97.5%
#> 1  4.6 0.185 4.6 4.24  4.97
#> pv
#>      mean   sd  50%   2.5% 97.5%
#> [1,] 4.61 2.74 4.63 -0.708    10
#> fd
#>     mean    sd    50%  2.5%  97.5%
#> 1 -0.833 0.235 -0.836 -1.29 -0.365