Method for plotting qi simulations across a range within a variable, with confidence intervals

Method for plotting qi simulations across a range within a variable, with confidence intervals

ci.plot(obj, qi="ev", var=NULL, ..., main = NULL, sub =
 NULL, xlab = NULL, ylab = NULL, xlim = NULL, ylim =
 NULL, legcol="gray20", col=NULL, leg=1, legpos=
 NULL, ci = c(80, 95, 99.9), discont=NULL)

Arguments

obj

A reference class zelig5 object

qi

a character-string specifying the quantity of interest to plot

var

The variable to be used on the x-axis. Default is the variable across all the chosen values with smallest nonzero variance

...

Parameters to be passed to the `truehist' function which is implicitly called for numeric simulations

main

a character-string specifying the main heading of the plot

sub

a character-string specifying the sub heading of the plot

xlab

a character-string specifying the label for the x-axis

ylab

a character-string specifying the label for the y-axis

xlim

Limits to the x-axis

ylim

Limits to the y-axis

legcol

``legend color'', an valid color used for plotting the line colors in the legend

col

a valid vector of colors of at least length 3 to use to color the confidence intervals

leg

``legend position'', an integer from 1 to 4, specifying the position of the legend. 1 to 4 correspond to ``SE'', ``SW'', ``NW'', and ``NE'' respectively. Setting to 0 or ``n'' turns off the legend.

legpos

``legend type'', exact coordinates and sizes for legend. Overrides argment ``leg.type''

ci

vector of length three of confidence interval levels to draw.

discont

optional point of discontinuity along the x-axis at which to interupt the graph

Value

the current graphical parameters. This is subject to change in future implementations of Zelig