Combines estimated coefficients and associated statistics

Combines estimated coefficients and associated statistics from models estimated with multiply imputed data sets or bootstrapped

combine_coef_se(obj, bagging = FALSE, messages = TRUE)

Arguments

obj

a zelig object with an estimated model

bagging

logical whether or not to bag the bootstrapped coefficients

messages

logical whether or not to return messages for what is being returned

Source

Partially based on mi.meld from Amelia.

Value

If the model uses multiply imputed or bootstrapped data then a list of combined coefficients (coef), standard errors (se), z values (zvalue), p-values (p) is returned. Rubin's Rules are used to combine output from multiply imputed data. An error is returned if no imputations were included or there wasn't bootstrapping. Please use get_coef, get_se, and get_pvalue methods instead in cases where there are no imputations or bootstrap.

Examples

set.seed(123)

## Multiple imputation example
# Create fake imputed data
n <- 100
x1 <- runif(n)
x2 <- runif(n)
y <- rnorm(n)
data.1 <- data.frame(y = y, x = x1)
data.2 <- data.frame(y = y, x = x2)

# Estimate model
mi.out <- to_zelig_mi(data.1, data.2)
z.out.mi <- zelig(y ~ x, model = "ls", data = mi.out)

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

# Combine and extract coefficients and standard errors
combine_coef_se(z.out.mi)

#> Combining imputations. . .

#> $coef
#> (Intercept)           x 
#> -0.12240923  0.03042057 
#> 
#> $se
#> (Intercept)           x 
#>   0.2139979   0.3763320 
#> 
#> $zvalue
#> (Intercept)           x 
#> -0.57201129  0.08083441 
#> 
#> $p
#> (Intercept)           x 
#>   0.5673143   0.9355736 
#> 

## Bootstrap example
z.out.boot <- zelig(y ~ x, model = "ls", data = data.1, bootstrap = 20)

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

combine_coef_se(z.out.boot)

#> Combining bootstraps . . .

#> $coef
#> (Intercept)           x 
#> -0.15723024  0.09965409 
#> 
#> $se
#> (Intercept)           x 
#>   0.2369467   0.3852040 
#> 
#> $zvalue
#> (Intercept)           x 
#>  -0.6635679   0.2587048 
#> 
#> $p
#> (Intercept)           x 
#>   0.5069669   0.7958631 
#>