Method for extracting estimated fitted values from Zelig objects

Method for extracting estimated fitted values from Zelig objects

# S4 method for Zelig
fitted(object, ...)

Arguments

object
An Object of Class Zelig

...
Additional parameters to be passed to fitted