Left Party Seat Share in 11 OECD Countries

This data set contains time-series data of the seat shares in the lower legislative house of left leaning parties over time, as well as the level of unemployment. Data follows the style used in Hibbs (1977).

data(seatshare)

Format

A table containing N variables ("country","year","unemp","leftseat") and 384 observations split across 11 countries.

Source

OECD data and Mackie and Rose (1991), extended to further years.

References

Douglas A. Hibbs. (1977). Political Parties and Macroeconomic Policy. American Political Science Review 71(4):1467-1487.

Thomas T. Mackie and Richard Rose. (1991). The International Almanac of Electoral History Macmillan: London.