Estimating a Statistical Model

The zelig function estimates a variety of statistical models. Use zelig output with setx and sim to compute quantities of interest, such as predicted probabilities, expected values, and first differences, along with the associated measures of uncertainty (standard errors and confidence intervals).

zelig(formula, model, data, ..., by = NULL, cite = TRUE)

Arguments

formula

a symbolic representation of the model to be estimated, in the form y \~\, x1 + x2, where y is the dependent variable and x1 and x2 are the explanatory variables, and y, x1, and x2 are contained in the same dataset. (You may include more than two explanatory variables, of course.) The + symbol means ``inclusion'' not ``addition.'' You may also include interaction terms and main effects in the form x1*x2 without computing them in prior steps; I(x1*x2) to include only the interaction term and exclude the main effects; and quadratic terms in the form I(x1^2)

model

the name of a statistical model. For a list of supported models visit: http://docs.zeligproject.org/en/latest/.

data

the name of a data frame containing the variables referenced in the formula, or a list of multiply imputed data frames each having the same variable names and row numbers (created by mi)

...

additional arguments passed to zelig, depending on the model to be estimated

by

a factor variable contained in data. Zelig will subset the data frame based on the levels in the by variable, and estimate a model for each subset. This a particularly powerful option which will allow you to save a considerable amount of effort. For example, to run the same model on all fifty states, you could type: z.out <- zelig(y ~ x1 + x2, data = mydata, model = 'ls', by = 'state') You may also use by to run models using MatchIt subclass

cite

If is set to 'TRUE' (default), the model citation will be

Value

Depending on the class of model selected, zelig will return an object with elements including coefficients, residuals, and formula which may be summarized using summary(z.out) or individually extracted using, for example, z.out\$coefficients. See the specific models listed above for additional output values, or simply type names(z.out).

Details

This documentation describes the zelig Zelig 4 compatibility wrapper function.