Extracted fitted values from a Zelig object with

Extracted fitted values from a Zelig object with setx values

zelig_setx_to_df(obj)

Arguments

obj

a zelig object with simulated quantities of interest

Details

Fitted (setx) values in a tidy data formatted data.frame. This was designed to enable the WhatIf package's whatif function to extract "counterfactuals".

Examples

#### QIs without first difference or range, from covariates fitted at
## central tendencies
z.1 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
             model = "ls")

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

z.1 <- setx(z.1)
zelig_setx_to_df(z.1)

#>   Petal.Length   Species
#> 1        3.758 virginica

#### QIs for first differences
z.2 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
             model = "ls")

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

z.2 <- setx(z.2, Petal.Length = 2)
z.2 <- setx1(z.2, Petal.Length = 4.4)
zelig_setx_to_df(z.2)

#>    Petal.Length   Species
#> 1           2.0 virginica
#> 11          4.4 virginica

#### QIs for first differences, estimated by Species
z.3 <- zelig(Petal.Width ~ Petal.Length, by = "Species", data = iris,
             model = "ls")

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

z.3 <- setx(z.3, Petal.Length = 2)
z.3 <- setx1(z.3, Petal.Length = 4.4)
zelig_setx_to_df(z.3)

#>           by Petal.Length
#> 1     setosa          2.0
#> 2 versicolor          2.0
#> 3  virginica          2.0
#> 4     setosa          4.4
#> 5 versicolor          4.4
#> 6  virginica          4.4

#### QIs for a range of fitted values
z.4 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
             model = "ls")

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

z.4 <- setx(z.4, Petal.Length = 2:4)
zelig_setx_to_df(z.4)

#>    Petal.Length   Species
#> 1             2 virginica
#> 11            3 virginica
#> 12            4 virginica

#### QIs for a range of fitted values, estimated by Species
z.5 <- zelig(Petal.Width ~ Petal.Length, by = "Species", data = iris,
             model = "ls")

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

z.5 <- setx(z.5, Petal.Length = 2:4)
zelig_setx_to_df(z.5)

#>           by Petal.Length
#> 1     setosa            2
#> 2 versicolor            2
#> 3  virginica            2
#> 4     setosa            3
#> 5 versicolor            3
#> 6  virginica            3
#> 7     setosa            4
#> 8 versicolor            4
#> 9  virginica            4

#### QIs for two ranges of fitted values
z.6 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
             model = "ls")

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

z.6 <- setx(z.6, Petal.Length = 2:4, Species = "setosa")
z.6 <- setx1(z.6, Petal.Length = 2:4, Species = "virginica")
zelig_setx_to_df(z.6)

#>     Petal.Length   Species
#> 1              2    setosa
#> 11             3    setosa
#> 12             4    setosa
#> 13             2 virginica
#> 111            3 virginica
#> 121            4 virginica