Generic Method for Computing and Organizing Simulated Quantities of Interest

Simulate quantities of interest from the estimated model output from zelig() given specified values of explanatory variables established in setx(). For classical maximum likelihood models, sim() uses asymptotic normal approximation to the log-likelihood. For Bayesian models, Zelig simulates quantities of interest from the posterior density, whenever possible. For robust Bayesian models, simulations are drawn from the identified class of Bayesian posteriors. Alternatively, you may generate quantities of interest using bootstrapped parameters.

sim(obj, x, x1, y = NULL, num = 1000, bootstrap = F, bootfn = NULL,
  cond.data = NULL, ...)

Arguments

obj

output object from zelig

x

values of explanatory variables used for simulation, generated by setx. Not if ommitted, then sim will look for values in the reference class object

x1

optional values of explanatory variables (generated by a second call of setx) particular computations of quantities of interest

y

a parameter reserved for the computation of particular quantities of interest (average treatment effects). Few models currently support this parameter

num

an integer specifying the number of simulations to compute

bootstrap

currently unsupported

bootfn

currently unsupported

cond.data

currently unsupported

...

arguments reserved future versions of Zelig

Value

The output stored in s.out varies by model. Use the names function to view the output stored in s.out. Common elements include:

Details

This documentation describes the sim Zelig 4 compatibility wrapper function.