Bundle Multiply Imputed Data Sets into an Object for Zelig

This object prepares multiply imputed data sets so they can be used by zelig.

to_zelig_mi(...)

Arguments

...

a set of data.frame's or a single list of data.frame's

Value

an mi object composed of a list of data frames.

Note

This function creates a list of data.frame objects, which resembles the storage of imputed data sets in the amelia object.

Examples

# create datasets
n <- 100
x1 <- runif(n)
x2 <- runif(n)
y <- rnorm(n)
data.1 <- data.frame(y = y, x = x1)
data.2 <- data.frame(y = y, x = x2)

# merge datasets into one object as if imputed datasets

mi.out <- to_zelig_mi(data.1, data.2)

# pass object in place of data argument
z.out <- zelig(y ~ x, model = "ls", data = mi.out)

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/