Extract simulated quantities of interest from a zelig object

Extract simulated quantities of interest from a zelig object

zelig_qi_to_df(obj)

Arguments

obj

a zelig object with simulated quantities of interest

Source

For a discussion of tidy data see https://vita.had.co.nz/papers/tidy-data.pdf.

Details

A simulated quantities of interest in a tidy data formatted data.frame. This can be useful for creating custom plots.

Each row contains a simulated value and each column contains:

setx_value whether the simulations are from the base x setx or the contrasting x1 for finding first differences.

The fitted values specified in setx including a by column if by was used in the zelig call.

expected_value

expected_value

See also

qi_slimmer

Examples

#### QIs without first difference or range, from covariates fitted at
## central tendencies
z.1 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
             model = "ls")

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

z.1 <- setx(z.1)
z.1 <- sim(z.1)
head(zelig_qi_to_df(z.1))

#> zelig_qi_to_df is an experimental function.
#>   Please report issues to: https://github.com/IQSS/Zelig/issues

#>   setx_value Petal.Length Speciesversicolor Speciesvirginica expected_value
#> 1          x        3.758                 0                1       1.610558
#> 2          x        3.758                 0                1       1.593792
#> 3          x        3.758                 0                1       1.586471
#> 4          x        3.758                 0                1       1.745465
#> 5          x        3.758                 0                1       1.569805
#> 6          x        3.758                 0                1       1.728413
#>   predicted_value
#> 1        1.688524
#> 2        1.484323
#> 3        1.898669
#> 4        1.776039
#> 5        1.460228
#> 6        1.569440

#### QIs for first differences
z.2 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
             model = "ls")

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

z.2a <- setx(z.2, Petal.Length = 2)
z.2b <- setx(z.2, Petal.Length = 4.4)
z.2 <- sim(z.2, x = z.2a, x1 = z.2a)
head(zelig_qi_to_df(z.2))

#> zelig_qi_to_df is an experimental function.
#>   Please report issues to: https://github.com/IQSS/Zelig/issues

#>   setx_value Petal.Length Speciesversicolor Speciesvirginica expected_value
#> 1          x            2                 0                1       1.105795
#> 2          x            2                 0                1       1.295009
#> 3          x            2                 0                1       1.360222
#> 4          x            2                 0                1       1.111205
#> 5          x            2                 0                1       1.189240
#> 6          x            2                 0                1       1.091091
#>   predicted_value
#> 1       1.6319970
#> 2       1.2442379
#> 3       1.4357070
#> 4       1.4465414
#> 5       1.3705797
#> 6       0.7106396

#### QIs for first differences, estimated by Species
z.3 <- zelig(Petal.Width ~ Petal.Length, by = "Species", data = iris,
             model = "ls")

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

z.3a <- setx(z.3, Petal.Length = 2)
z.3b <- setx(z.3, Petal.Length = 4.4)
z.3 <- sim(z.3, x = z.3a, x1 = z.3a)
head(zelig_qi_to_df(z.3))

#> zelig_qi_to_df is an experimental function.
#>   Please report issues to: https://github.com/IQSS/Zelig/issues

#>   setx_value     by Petal.Length expected_value predicted_value
#> 1          x setosa            2      0.3168079       0.4511546
#> 2          x setosa            2      0.3962924       0.3595626
#> 3          x setosa            2      0.3328195       0.3280413
#> 4          x setosa            2      0.3807603       0.4341890
#> 5          x setosa            2      0.3478657       0.2472453
#> 6          x setosa            2      0.3338638       0.2325748

#### QIs for a range of fitted values
z.4 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
             model = "ls")

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

z.4 <- setx(z.4, Petal.Length = 2:4)
z.4 <- sim(z.4)
head(zelig_qi_to_df(z.4))

#> zelig_qi_to_df is an experimental function.
#>   Please report issues to: https://github.com/IQSS/Zelig/issues

#>   setx_value Petal.Length Speciesversicolor Speciesvirginica expected_value
#> 1          x            2                 0                1      1.3436809
#> 2          x            2                 0                1      1.0169605
#> 3          x            2                 0                1      0.9829203
#> 4          x            2                 0                1      0.9617881
#> 5          x            2                 0                1      1.3278357
#> 6          x            2                 0                1      0.9850973
#>   predicted_value
#> 1       1.5409543
#> 2       1.0849553
#> 3       0.8413208
#> 4       1.3057269
#> 5       1.3640283
#> 6       0.9146892

#### QIs for a range of fitted values, estimated by Species
z.5 <- zelig(Petal.Width ~ Petal.Length, by = "Species", data = iris,
            model = "ls")

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

z.5 <- setx(z.5, Petal.Length = 2:4)
z.5 <- sim(z.5)
head(zelig_qi_to_df(z.5))

#> zelig_qi_to_df is an experimental function.
#>   Please report issues to: https://github.com/IQSS/Zelig/issues

#>   setx_value     by Petal.Length expected_value predicted_value
#> 1          x setosa            2      0.3223436       0.1491181
#> 2          x setosa            2      0.3574439       0.4088742
#> 3          x setosa            2      0.3141658       0.2946930
#> 4          x setosa            2      0.3652069       0.2462607
#> 5          x setosa            2      0.3506836       0.3749823
#> 6          x setosa            2      0.3850459       0.5033220

#### QIs for two ranges of fitted values
z.6 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
            model = "ls")

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

z.6a <- setx(z.6, Petal.Length = 2:4, Species = "setosa")
z.6b <- setx(z.6, Petal.Length = 2:4, Species = "virginica")
z.6 <- sim(z.6, x = z.6a, x1 = z.6b)

head(zelig_qi_to_df(z.6))

#> zelig_qi_to_df is an experimental function.
#>   Please report issues to: https://github.com/IQSS/Zelig/issues

#>   setx_value Petal.Length Speciesversicolor Speciesvirginica expected_value
#> 1          x            2                 0                0      0.3879272
#> 2          x            2                 0                0      0.3548106
#> 3          x            2                 0                0      0.3784270
#> 4          x            2                 0                0      0.3777677
#> 5          x            2                 0                0      0.3483144
#> 6          x            2                 0                0      0.3501803
#>   predicted_value
#> 1     -0.02117684
#> 2      0.63640797
#> 3      0.64356316
#> 4      0.58384380
#> 5      0.33351132
#> 6      0.05126477