Extract simulated quantities of interest from a zelig object

Extract simulated quantities of interest from a zelig object

zelig_qi_to_df(obj)

Arguments

obj
a zelig object with simulated quantities of interest

Source

For a discussion of tidy data see https://www.jstatsoft.org/article/view/v059i10.

Details

A simulated quantities of interest in a tidy data formatted data.frame. This can be useful for creating custom plots.

Each row contains a simulated value and each column contains:

setx_value whether the simulations are from the base x setx or the contrasting x1 for finding first differences.

The fitted values specified in setx including a by column if by was used in the zelig call.

expected_value

predicted_value

For multinomial reponse models, a separate column is given for the expected probability of each outcome in the form expected_*. Additionally, there a is column of the predicted outcomes (predicted_value).

See also

qi_slimmer

Examples

#### QIs without first difference or range, from covariates fitted at
## central tendencies
z.1 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
             model = "ls")

#> Error in Petal.Width ~ Petal.Length + Species: unused arguments (formula = Petal.Width, data = Petal.Length + Species)

z.1 <- setx(z.1)

#> Error in inherits(x, "Zelig"): object 'z.1' not found

z.1 <- sim(z.1)

#> Error in inherits(x, "Zelig"): object 'z.1' not found

head(zelig_qi_to_df(z.1))

#> Error in inherits(x, "Zelig"): object 'z.1' not found

#### QIs for first differences
z.2 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
             model = "ls")

#> Error in Petal.Width ~ Petal.Length + Species: unused arguments (formula = Petal.Width, data = Petal.Length + Species)

z.2a <- setx(z.2, Petal.Length = 2)

#> Error in inherits(x, "Zelig"): object 'z.2' not found

z.2b <- setx(z.2, Petal.Length = 4.4)

#> Error in inherits(x, "Zelig"): object 'z.2' not found

z.2 <- sim(z.2, x = z.2a, x1 = z.2a)

#> Error in inherits(x, "Zelig"): object 'z.2' not found

head(zelig_qi_to_df(z.2))

#> Error in inherits(x, "Zelig"): object 'z.2' not found

#### QIs for first differences, estimated by Species
z.3 <- zelig(Petal.Width ~ Petal.Length, by = "Species", data = iris,
             model = "ls")

#> Error in Petal.Width ~ Petal.Length: unused arguments (formula = Petal.Width, data = Petal.Length)

z.3a <- setx(z.3, Petal.Length = 2)

#> Error in inherits(x, "Zelig"): object 'z.3' not found

z.3b <- setx(z.3, Petal.Length = 4.4)

#> Error in inherits(x, "Zelig"): object 'z.3' not found

z.3 <- sim(z.3, x = z.3a, x1 = z.3a)

#> Error in inherits(x, "Zelig"): object 'z.3' not found

head(zelig_qi_to_df(z.3))

#> Error in inherits(x, "Zelig"): object 'z.3' not found

#### QIs for a range of fitted values
z.4 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
             model = "ls")

#> Error in Petal.Width ~ Petal.Length + Species: unused arguments (formula = Petal.Width, data = Petal.Length + Species)

z.4 <- setx(z.4, Petal.Length = 2:4)

#> Error in inherits(x, "Zelig"): object 'z.4' not found

z.4 <- sim(z.4)

#> Error in inherits(x, "Zelig"): object 'z.4' not found

head(zelig_qi_to_df(z.4))

#> Error in inherits(x, "Zelig"): object 'z.4' not found

#### QIs for a range of fitted values, estimated by Species
z.5 <- zelig(Petal.Width ~ Petal.Length, by = "Species", data = iris,
            model = "ls")

#> Error in Petal.Width ~ Petal.Length: unused arguments (formula = Petal.Width, data = Petal.Length)

z.5 <- setx(z.5, Petal.Length = 2:4)

#> Error in inherits(x, "Zelig"): object 'z.5' not found

z.5 <- sim(z.5)

#> Error in inherits(x, "Zelig"): object 'z.5' not found

head(zelig_qi_to_df(z.5))

#> Error in inherits(x, "Zelig"): object 'z.5' not found

#### QIs for two ranges of fitted values
z.6 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
            model = "ls")

#> Error in Petal.Width ~ Petal.Length + Species: unused arguments (formula = Petal.Width, data = Petal.Length + Species)

z.6a <- setx(z.6, Petal.Length = 2:4, Species = "setosa")

#> Error in inherits(x, "Zelig"): object 'z.6' not found

z.6b <- setx(z.6, Petal.Length = 2:4, Species = "virginica")

#> Error in inherits(x, "Zelig"): object 'z.6' not found

z.6 <- sim(z.6, x = z.6a, x1 = z.6b)

#> Error in inherits(x, "Zelig"): object 'z.6' not found

head(zelig_qi_to_df(z.6))

#> Error in inherits(x, "Zelig"): object 'z.6' not found