Extract simulated quantities of interest from a zelig object

Extract simulated quantities of interest from a zelig object

zelig_qi_to_df(obj)

Arguments

obj

a zelig object with simulated quantities of interest

Source

For a discussion of tidy data see https://vita.had.co.nz/papers/tidy-data.pdf.

Details

A simulated quantities of interest in a tidy data formatted data.frame. This can be useful for creating custom plots.

Each row contains a simulated value and each column contains:

setx_value whether the simulations are from the base x setx or the contrasting x1 for finding first differences.

The fitted values specified in setx including a by column if by was used in the zelig call.

expected_value

expected_value

See also

qi_slimmer

Examples

#### QIs without first difference or range, from covariates fitted at
## central tendencies
z.1 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
             model = "ls")

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

z.1 <- setx(z.1)
z.1 <- sim(z.1)
head(zelig_qi_to_df(z.1))

#> zelig_qi_to_df is an experimental function.
#>   Please report issues to: https://github.com/IQSS/Zelig/issues

#>   setx_value Petal.Length Speciesversicolor Speciesvirginica expected_value
#> 1          x        3.758                 0                0      0.8425537
#> 2          x        3.758                 0                0      0.7301953
#> 3          x        3.758                 0                0      0.8379955
#> 4          x        3.758                 0                0      0.7453869
#> 5          x        3.758                 0                0      0.5527217
#> 6          x        3.758                 0                0      0.7656386
#>   predicted_value
#> 1       0.6606115
#> 2       0.7980091
#> 3       0.5025046
#> 4       0.8961028
#> 5       0.4925887
#> 6       1.1736058

#### QIs for first differences
z.2 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
             model = "ls")

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

z.2a <- setx(z.2, Petal.Length = 2)
z.2b <- setx(z.2, Petal.Length = 4.4)
z.2 <- sim(z.2, x = z.2a, x1 = z.2a)
head(zelig_qi_to_df(z.2))

#> zelig_qi_to_df is an experimental function.
#>   Please report issues to: https://github.com/IQSS/Zelig/issues

#>   setx_value Petal.Length Speciesversicolor Speciesvirginica expected_value
#> 1          x            2                 1                0      0.8907621
#> 2          x            2                 1                0      0.7144591
#> 3          x            2                 1                0      0.8322978
#> 4          x            2                 1                0      0.7542331
#> 5          x            2                 1                0      0.6290049
#> 6          x            2                 1                0      0.8487060
#>   predicted_value
#> 1       0.6877237
#> 2       0.6865897
#> 3       0.9772715
#> 4       0.9091981
#> 5       0.8436772
#> 6       0.8599703

#### QIs for first differences, estimated by Species
z.3 <- zelig(Petal.Width ~ Petal.Length, by = "Species", data = iris,
             model = "ls")

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

z.3a <- setx(z.3, Petal.Length = 2)
z.3b <- setx(z.3, Petal.Length = 4.4)
z.3 <- sim(z.3, x = z.3a, x1 = z.3a)
head(zelig_qi_to_df(z.3))

#> zelig_qi_to_df is an experimental function.
#>   Please report issues to: https://github.com/IQSS/Zelig/issues

#>   setx_value     by Petal.Length expected_value predicted_value
#> 1          x setosa            2      0.3699721       0.3489035
#> 2          x setosa            2      0.4045550       0.3198680
#> 3          x setosa            2      0.3643632       0.3865619
#> 4          x setosa            2      0.3456148       0.3936453
#> 5          x setosa            2      0.2890379       0.1255503
#> 6          x setosa            2      0.3483292       0.5026930

#### QIs for a range of fitted values
z.4 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
             model = "ls")

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

z.4 <- setx(z.4, Petal.Length = 2:4)
z.4 <- sim(z.4)
head(zelig_qi_to_df(z.4))

#> zelig_qi_to_df is an experimental function.
#>   Please report issues to: https://github.com/IQSS/Zelig/issues

#>   setx_value Petal.Length Speciesversicolor Speciesvirginica expected_value
#> 1          x            2                 0                1      0.9850973
#> 2          x            2                 0                1      1.1994781
#> 3          x            2                 0                1      1.3213836
#> 4          x            2                 0                1      1.0466962
#> 5          x            2                 0                1      1.1627801
#> 6          x            2                 0                1      1.0586074
#>   predicted_value
#> 1       0.9146892
#> 2       1.0549579
#> 3       1.2872477
#> 4       1.2370018
#> 5       1.4929164
#> 6       1.1477272

#### QIs for a range of fitted values, estimated by Species
z.5 <- zelig(Petal.Width ~ Petal.Length, by = "Species", data = iris,
            model = "ls")

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

z.5 <- setx(z.5, Petal.Length = 2:4)
z.5 <- sim(z.5)
head(zelig_qi_to_df(z.5))

#> zelig_qi_to_df is an experimental function.
#>   Please report issues to: https://github.com/IQSS/Zelig/issues

#>   setx_value     by Petal.Length expected_value predicted_value
#> 1          x setosa            2      0.3850459       0.5033220
#> 2          x setosa            2      0.2248206       0.2588107
#> 3          x setosa            2      0.3432744       0.3803239
#> 4          x setosa            2      0.3380666       0.2859084
#> 5          x setosa            2      0.3942121       0.4235448
#> 6          x setosa            2      0.4716071       0.4122104

#### QIs for two ranges of fitted values
z.6 <- zelig(Petal.Width ~ Petal.Length + Species, data = iris,
            model = "ls")

#> How to cite this model in Zelig:
#>   R Core Team. 2007.
#>   ls: Least Squares Regression for Continuous Dependent Variables
#>   in Christine Choirat, Christopher Gandrud, James Honaker, Kosuke Imai, Gary King, and Olivia Lau,
#>   "Zelig: Everyone's Statistical Software," http://zeligproject.org/

z.6a <- setx(z.6, Petal.Length = 2:4, Species = "setosa")
z.6b <- setx(z.6, Petal.Length = 2:4, Species = "virginica")
z.6 <- sim(z.6, x = z.6a, x1 = z.6b)

head(zelig_qi_to_df(z.6))

#> zelig_qi_to_df is an experimental function.
#>   Please report issues to: https://github.com/IQSS/Zelig/issues

#>   setx_value Petal.Length Speciesversicolor Speciesvirginica expected_value
#> 1          x            2                 0                0      0.3993282
#> 2          x            2                 0                0      0.3542462
#> 3          x            2                 0                0      0.3822508
#> 4          x            2                 0                0      0.3606174
#> 5          x            2                 0                0      0.3613591
#> 6          x            2                 0                0      0.3707449
#>   predicted_value
#> 1      0.45451559
#> 2      0.22340240
#> 3      0.33441671
#> 4      0.27230319
#> 5      0.02313176
#> 6      0.35664633